chore: formatting

+++ libs/coupldyn_yac/yac_cartesian_dynamics.cpp
@@ -9,7 +9,7 @@
  * Author: Clara Bayley (CB)
  * Additional Contributors:
  * -----
- * Last Modified: Wednesday 17th April 2024
+ * Last Modified: Tuesday 18th June 2024
  * Modified By: CB
  * -----
  * License: BSD 3-Clause "New" or "Revised" License
@@ -23,9 +23,9 @@
 
 #include "coupldyn_yac/yac_cartesian_dynamics.hpp"
 
-#include <iostream>
-
 #include <mpi.h>
+
+#include <iostream>
 extern "C" {
 #include "yac.h"
 }
@@ -43,10 +43,9 @@ std::array<size_t, 3> kijfromindex(const std::array<size_t, 3> &ndims, const siz
   return std::array<size_t, 3>{k, i, j};
 }
 
-void create_vertex_coordinates(const Config &config,
-                               const std::array<size_t, 3> ndims,
-                               std::vector<double> & vertex_longitudes,
-                               std::vector<double> & vertex_latitudes) {
+void create_vertex_coordinates(const Config &config, const std::array<size_t, 3> ndims,
+                               std::vector<double> &vertex_longitudes,
+                               std::vector<double> &vertex_latitudes) {
   double lower_longitude = 0;
   double upper_longitude = ndims[0] * (2 * std::numbers::pi / (ndims[0] + 1));
   double lower_latitude = (-0.5 * std::numbers::pi * ndims[1]) / (ndims[1] + 2);
@@ -56,10 +55,10 @@ void create_vertex_coordinates(const Config &config,
       config.get_yac_dynamics().upper_longitude != NaNVals::dbl() &&
       config.get_yac_dynamics().lower_latitude != NaNVals::dbl() &&
       config.get_yac_dynamics().upper_latitude != NaNVals::dbl()) {
-      lower_longitude = config.get_yac_dynamics().lower_longitude;
-      upper_longitude = config.get_yac_dynamics().upper_longitude;
-      lower_latitude = config.get_yac_dynamics().lower_latitude;
-      upper_latitude = config.get_yac_dynamics().upper_latitude;
+    lower_longitude = config.get_yac_dynamics().lower_longitude;
+    upper_longitude = config.get_yac_dynamics().upper_longitude;
+    lower_latitude = config.get_yac_dynamics().lower_latitude;
+    upper_latitude = config.get_yac_dynamics().upper_latitude;
   }
 
   // Defines the vertex longitude and latitude values in radians for grid creation
@@ -72,10 +71,9 @@ void create_vertex_coordinates(const Config &config,
 }
 
 /* Creates the YAC grid and defines the cell and edge points based on ndims data */
-void create_grid_and_points_definitions(const Config &config,
-                                        const std::array<size_t, 3> ndims,
-                                        const std::string grid_name,
-                                        int & grid_id, int & cell_point_id, int & edge_point_id) {
+void create_grid_and_points_definitions(const Config &config, const std::array<size_t, 3> ndims,
+                                        const std::string grid_name, int &grid_id,
+                                        int &cell_point_id, int &edge_point_id) {
   int cyclic_dimension[2] = {0, 0};
   int total_cells[2] = {static_cast<int>(ndims[0]), static_cast<int>(ndims[1])};
   int total_vertices[2] = {static_cast<int>(ndims[0] + 1), static_cast<int>(ndims[1] + 1)};
@@ -92,19 +90,19 @@ void create_grid_and_points_definitions(const Config &config,
   create_vertex_coordinates(config, ndims, vertex_longitudes, vertex_latitudes);
 
   // Defines a regular 2D grid
-  yac_cdef_grid_reg2d(grid_name.c_str(), total_vertices, cyclic_dimension,
-                      vertex_longitudes.data(), vertex_latitudes.data(), &grid_id);
+  yac_cdef_grid_reg2d(grid_name.c_str(), total_vertices, cyclic_dimension, vertex_longitudes.data(),
+                      vertex_latitudes.data(), &grid_id);
 
   // --- Point definitions ---
   // Defines the cell center longitude and latitude values in radians
   // The values are later permuted by YAC to generate all cell center coordinates
   for (size_t i = 0; i < cell_center_longitudes.size(); i++)
-    cell_center_longitudes[i] = vertex_longitudes[i]
-                                + (vertex_longitudes[i + 1] - vertex_longitudes[i]) / 2;
+    cell_center_longitudes[i] =
+        vertex_longitudes[i] + (vertex_longitudes[i + 1] - vertex_longitudes[i]) / 2;
 
   for (size_t i = 0; i < cell_center_latitudes.size(); i++)
-    cell_center_latitudes[i] = vertex_latitudes[i]
-                               + (vertex_latitudes[i + 1] - vertex_latitudes[i]) / 2;
+    cell_center_latitudes[i] =
+        vertex_latitudes[i] + (vertex_latitudes[i + 1] - vertex_latitudes[i]) / 2;
 
   // Defines the edge center longitude and latitude values in radians.
   // Since it is not possible to generate edge center coordinates with a single
@@ -133,13 +131,11 @@ void create_grid_and_points_definitions(const Config &config,
                            &edge_point_id);
 }
 
-void CartesianDynamics::receive_yac_field(unsigned int field_type,
-                                          unsigned int yac_field_id,
-                                          double ** yac_raw_data,
-                                          std::vector<double> & target_array,
+void CartesianDynamics::receive_yac_field(unsigned int field_type, unsigned int yac_field_id,
+                                          double **yac_raw_data, std::vector<double> &target_array,
                                           size_t vertical_levels) {
   int info, error;
-  unsigned int total_horizontal_cells = ndims[0] * ndims[1];
+  size_t total_horizontal_cells = ndims[0] * ndims[1];
   bool edge_dimension = false;
   std::vector<double>::iterator target_it = target_array.begin();
 
@@ -177,11 +173,7 @@ void CartesianDynamics::receive_yac_field(unsigned int field_type,
 /* This subroutine is the main entry point for receiving data from YAC.
  * It checks the dimensionality of the simulation based on the config data. */
 void CartesianDynamics::receive_field_collections_from_yac() {
-  enum field_types {
-    CELL,
-    U_EDGE,
-    W_EDGE
-  };
+  enum field_types { CELL, U_EDGE, W_EDGE };
 
   receive_yac_field(CELL, temp_yac_id, yac_raw_cell_data, temp, ndims[2]);
   receive_yac_field(CELL, pressure_yac_id, yac_raw_cell_data, press, ndims[2]);
@@ -219,8 +211,8 @@ CartesianDynamics::CartesianDynamics(const Config &config, const std::array<size
   int edge_point_id = -1;
   std::string grid_name = "cleo_grid";
 
-  create_grid_and_points_definitions(config, ndims, grid_name, grid_id,
-                                     cell_point_id, edge_point_id);
+  create_grid_and_points_definitions(config, ndims, grid_name, grid_id, cell_point_id,
+                                     edge_point_id);
 
   // --- Interpolation stack ---
   int interp_stack_id;
@@ -232,33 +224,30 @@ CartesianDynamics::CartesianDynamics(const Config &config, const std::array<size
   int horizontal_fields_collection_size = ndims[2];
   int vertical_winds_collection_size = ndims[2] + 1;
 
-  yac_cdef_field("pressure", component_id, &cell_point_id,
-                 num_point_sets, horizontal_fields_collection_size, "PT1M",
-                 YAC_TIME_UNIT_ISO_FORMAT, &pressure_yac_id);
+  yac_cdef_field("pressure", component_id, &cell_point_id, num_point_sets,
+                 horizontal_fields_collection_size, "PT1M", YAC_TIME_UNIT_ISO_FORMAT,
+                 &pressure_yac_id);
 
-  yac_cdef_field("temperature", component_id, &cell_point_id,
-                 num_point_sets, horizontal_fields_collection_size, "PT1M",
-                 YAC_TIME_UNIT_ISO_FORMAT, &temp_yac_id);
+  yac_cdef_field("temperature", component_id, &cell_point_id, num_point_sets,
+                 horizontal_fields_collection_size, "PT1M", YAC_TIME_UNIT_ISO_FORMAT, &temp_yac_id);
 
-  yac_cdef_field("qvap", component_id, &cell_point_id,
-                 num_point_sets, horizontal_fields_collection_size, "PT1M",
-                 YAC_TIME_UNIT_ISO_FORMAT, &qvap_yac_id);
+  yac_cdef_field("qvap", component_id, &cell_point_id, num_point_sets,
+                 horizontal_fields_collection_size, "PT1M", YAC_TIME_UNIT_ISO_FORMAT, &qvap_yac_id);
 
-  yac_cdef_field("qcond", component_id, &cell_point_id,
-                 num_point_sets, horizontal_fields_collection_size, "PT1M",
-                 YAC_TIME_UNIT_ISO_FORMAT, &qcond_yac_id);
+  yac_cdef_field("qcond", component_id, &cell_point_id, num_point_sets,
+                 horizontal_fields_collection_size, "PT1M", YAC_TIME_UNIT_ISO_FORMAT,
+                 &qcond_yac_id);
 
-  yac_cdef_field("eastward_wind", component_id, &edge_point_id,
-                 num_point_sets, horizontal_fields_collection_size, "PT1M",
-                 YAC_TIME_UNIT_ISO_FORMAT, &eastward_wind_yac_id);
+  yac_cdef_field("eastward_wind", component_id, &edge_point_id, num_point_sets,
+                 horizontal_fields_collection_size, "PT1M", YAC_TIME_UNIT_ISO_FORMAT,
+                 &eastward_wind_yac_id);
 
-  yac_cdef_field("northward_wind", component_id, &edge_point_id,
-                 num_point_sets, horizontal_fields_collection_size, "PT1M",
-                 YAC_TIME_UNIT_ISO_FORMAT, &northward_wind_yac_id);
+  yac_cdef_field("northward_wind", component_id, &edge_point_id, num_point_sets,
+                 horizontal_fields_collection_size, "PT1M", YAC_TIME_UNIT_ISO_FORMAT,
+                 &northward_wind_yac_id);
 
-  yac_cdef_field("vvel", component_id, &cell_point_id,
-                 num_point_sets, vertical_winds_collection_size, "PT1M",
-                 YAC_TIME_UNIT_ISO_FORMAT, &vvel_yac_id);
+  yac_cdef_field("vvel", component_id, &cell_point_id, num_point_sets,
+                 vertical_winds_collection_size, "PT1M", YAC_TIME_UNIT_ISO_FORMAT, &vvel_yac_id);
 
   // --- Field coupling definitions ---
   yac_cdef_couple("icon_data_reader", "bubble_grid", "pressure", "cleo", "cleo_grid", "pressure",
@@ -274,13 +263,13 @@ CartesianDynamics::CartesianDynamics(const Config &config, const std::array<size
   yac_cdef_couple("icon_data_reader", "bubble_grid", "qcond", "cleo", "cleo_grid", "qcond", "PT1M",
                   YAC_TIME_UNIT_ISO_FORMAT, YAC_REDUCTION_TIME_NONE, interp_stack_id, 0, 0);
 
-  yac_cdef_couple("icon_data_reader", "bubble_grid", "eastward_wind",
-                  "cleo", "cleo_grid", "eastward_wind", "PT1M",
-                  YAC_TIME_UNIT_ISO_FORMAT, YAC_REDUCTION_TIME_NONE, interp_stack_id, 0, 0);
+  yac_cdef_couple("icon_data_reader", "bubble_grid", "eastward_wind", "cleo", "cleo_grid",
+                  "eastward_wind", "PT1M", YAC_TIME_UNIT_ISO_FORMAT, YAC_REDUCTION_TIME_NONE,
+                  interp_stack_id, 0, 0);
 
-  yac_cdef_couple("icon_data_reader", "bubble_grid", "northward_wind",
-                  "cleo", "cleo_grid", "northward_wind", "PT1M",
-                  YAC_TIME_UNIT_ISO_FORMAT, YAC_REDUCTION_TIME_NONE, interp_stack_id, 0, 0);
+  yac_cdef_couple("icon_data_reader", "bubble_grid", "northward_wind", "cleo", "cleo_grid",
+                  "northward_wind", "PT1M", YAC_TIME_UNIT_ISO_FORMAT, YAC_REDUCTION_TIME_NONE,
+                  interp_stack_id, 0, 0);
 
   yac_cdef_couple("icon_data_reader", "bubble_grid", "vvel", "cleo", "cleo_grid", "vvel", "PT1M",
                   YAC_TIME_UNIT_ISO_FORMAT, YAC_REDUCTION_TIME_NONE, interp_stack_id, 0, 0);
@@ -291,9 +280,9 @@ CartesianDynamics::CartesianDynamics(const Config &config, const std::array<size
   size_t horizontal_cell_number = yac_cget_grid_size(YAC_LOCATION_CELL, grid_id);
   size_t horizontal_edge_number = yac_cget_grid_size(YAC_LOCATION_EDGE, grid_id);
 
-  yac_raw_cell_data = new double * [ndims[2]];
-  yac_raw_edge_data = new double * [ndims[2]];
-  yac_raw_vertical_wind_data = new double * [ndims[2] + 1];
+  yac_raw_cell_data = new double *[ndims[2]];
+  yac_raw_edge_data = new double *[ndims[2]];
+  yac_raw_vertical_wind_data = new double *[ndims[2] + 1];
 
   for (size_t i = 0; i < ndims[2]; i++) {
     yac_raw_cell_data[i] = new double[horizontal_cell_number];
@@ -333,12 +322,11 @@ CartesianDynamics::~CartesianDynamics() {
     delete yac_raw_edge_data[i];
   }
 
-  for (size_t i = 0; i < ndims[2] + 1; i++)
-    delete yac_raw_vertical_wind_data[i];
+  for (size_t i = 0; i < ndims[2] + 1; i++) delete yac_raw_vertical_wind_data[i];
 
-  delete [] yac_raw_cell_data;
-  delete [] yac_raw_edge_data;
-  delete [] yac_raw_vertical_wind_data;
+  delete[] yac_raw_cell_data;
+  delete[] yac_raw_edge_data;
+  delete[] yac_raw_vertical_wind_data;
 
   yac_cfinalize();
 }
